Добавит компоненты элементов страниц.
