+++ src/components/map.tsx
@@ -1,0 +1,7 @@
+function MapComponent(): JSX.Element {
+  return(
+    <section className="cities__map map" />
+  );
+}
+
+export default MapComponent;
