+++ src/pages/main-screen/main-screen.tsx
@@ -1,50 +1,22 @@
-import PlaceCard from '../components/place-card';
+import CitiesCard from '../../components/cities-card';
+import Header from '../../components/header';
+import MapComponent from '../../components/map';
 
 type MainScreenProps = {
-  offersCount: number;
+  citiesCount: number;
 }
 
-function MainScreen({offersCount}: MainScreenProps): JSX.Element {
+function MainScreen({citiesCount}: MainScreenProps): JSX.Element {
+
+  let counter:number = 0;
+
+  const citiesCardArray: JSX.Element[] = Array.from({length: citiesCount}).map(() => <CitiesCard key={counter++} />);
+
   return (
     <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <a className="header__logo-link header__logo-link--active">
-                <img
-                  className="header__logo"
-                  src="img/logo.svg"
-                  alt="6 cities logo"
-                  width={81}
-                  height={41}
-                />
-              </a>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a
-                    className="header__nav-link header__nav-link--profile"
-                    href="#"
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    <span className="header__user-name user__name">
-                      Oliver.conner@gmail.com
-                    </span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+
+      <Header />
+
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
@@ -87,7 +59,7 @@ function MainScreen({offersCount}: MainScreenProps): JSX.Element {
           <div className="cities__places-container container">
             <section className="cities__places places">
               <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{offersCount} places to stay in Amsterdam</b>
+              <b className="places__found">{citiesCount} places to stay in Amsterdam</b>
               <form className="places__sorting" action="#" method="get">
                 <span className="places__sorting-caption">Sort by</span>
                 <span className="places__sorting-type" tabIndex={0}>
@@ -115,15 +87,11 @@ function MainScreen({offersCount}: MainScreenProps): JSX.Element {
                 </ul>
               </form>
               <div className="cities__places-list places__list tabs__content">
-                <PlaceCard />
-                <PlaceCard />
-                <PlaceCard />
-                <PlaceCard />
-                <PlaceCard />
+                {citiesCardArray}
               </div>
             </section>
             <div className="cities__right-section">
-              <section className="cities__map map" />
+              <MapComponent />
             </div>
           </div>
         </div>
